requirements.txt
flask
flask-cors
authlib
google-generativeai
python-dotenv
requests
werkzeug
